Split specs file into smaller files
